requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9

pillow>=10.0
numpy>=1.26

opencv-python-headless>=4.9
mediapipe>=0.10.14
ultralytics>=8.2

pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
